PHOENIX-5221 Phoenix Kerberos Integration tests failure on Redhat Linux
Due to this bug https://bugzilla.redhat.com/show_bug.cgi?id=668830
We need to use `localhost.localdomain` as host name when running these tests on Jenkins (Centos)
but for Mac OS it should be `localhost` to pass.
The reason is kerberos principals in this tests are looked up from /etc/hosts
and a reverse DNS lookup of 127.0.0.1 is resolved to `localhost.localdomain` rather than `localhost` on Centos.
KDC sees `localhost` != `localhost.localdomain` and as the result test fails with authentication error.
It's also important to note these principals are shared between HDFs and HBase in this mini HBase cluster.
Some more reading https://access.redhat.com/solutions/57330
